% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinet.r
\name{chinet}
\alias{chinet}
\title{chinet
perform network plot with lovain clustering on the results obtained by cltest function}
\usage{
chinet(
  results,
  fold = 1,
  cex = 2,
  distance = 3,
  family = "sans",
  layout = layout_as_star
)
}
\arguments{
\item{fold}{fold for edge weight variations}

\item{cex}{police size for labels of the vertex}

\item{distance}{distance between label and vertex}

\item{family}{family of font for the labels}

\item{layout}{layout style of the network (layout_as_star, layout_components, layout_in_circle, layout_nicely,layout_on_grid, layout_on_sphere, layout_randomly, layout_with_dh, layout_with_drl, layout_with_fr, layout_with_gem, layout_with_graphopt, layout_with_kk, layout_with_lgl, layout_with_mds)}
}
\description{
chinet
perform network plot with lovain clustering on the results obtained by cltest function
}
\examples{
data(carmodels)
df<-preprocess(carmodels)
results<-cltest(df)
chinet(results,fold=0.5,cex=2,distance=3,family="sans",layout=layout_in_circle)
}
