% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltest.r
\name{cltest}
\alias{cltest}
\title{cltest
loop chi.test by 2*2 matrix on character columns of a dataset}
\usage{
results<-cltest(models)
}
\description{
cltest
loop chi.test by 2*2 matrix on character columns of a dataset
}
\examples{
data(carmodels)
df<-preprocess(carmodels)
results<-cltest(df)
}
