% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpplot.r
\name{nlpplot}
\alias{nlpplot}
\title{nlpplot
perform matrix plot of the results obtained by cltest function}
\usage{
nlpplot(results, titleplot = "NLP plot", txt = 12)
}
\arguments{
\item{titleplot}{"string" for the title of the plot}

\item{txt}{size of the label for the title}
}
\description{
nlpplot
perform matrix plot of the results obtained by cltest function
}
\examples{
data(carmodels)
df<-preprocess(carmodels)
results<-cltest(df)
nlpplot(results,titleplot="NLP plot on carmodels",txt=12)
}
