% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.r
\name{preprocess}
\alias{preprocess}
\title{preprocess
this function take input dataframe with possible missing data and preprocessed it:
1- it transforms factor columns as character columns
2- it removes character columns having unique values}
\usage{
preprocess(carmodels)
}
\description{
preprocess
this function take input dataframe with possible missing data and preprocessed it:
1- it transforms factor columns as character columns
2- it removes character columns having unique values
}
\examples{
data(carmodels)
df<-preprocess(carmodels)
}
